Move DynamoDB management scripts to their own directory.
