+++ dynamodb/scripts/confirmCommand.js
